// Frontend/src/app/core/services/aminety-service/models/ReportSummaryDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ReportSummaryDto = {
    totalBookings?: number;
    totalPeople?: number;
    activeBookings?: number;
    cancelledBookings?: number;
    completedBookings?: number;
    utilizationRate?: number;
};